requirements.txt
pydantic==2.7.3
python-dotenv=1.0.0
pytelegrambotapi==4.17.0